+++ src/structlog/__init__.py
@@ -37,7 +37,7 @@ except ImportError:  # pragma: nocover
 else:
     twisted = _twisted
 
-__version__ = "19.2.0.dev0"
+__version__ = "19.3.0.dev0"
 
 __title__ = "structlog"
 __description__ = "Structured Logging for Python"