+++ src/structlog/_loggers.py
@@ -82,7 +82,7 @@ class PrintLogger(object):
     def __getstate__(self):
         # type: () -> str
         """
-        Out __getattr__ magic makes this necessary.
+        Our __getattr__ magic makes this necessary.
         """
         if self._file is sys.stdout:
             return "stdout"
@@ -97,7 +97,7 @@ class PrintLogger(object):
     def __setstate__(self, state):
         # type: (str) -> None
         """
-        Out __getattr__ magic makes this necessary.
+        Our __getattr__ magic makes this necessary.
         """
         if state == "stdout":
             self._file = sys.stdout