+++ src/structlog/dev.py
@@ -10,7 +10,7 @@ from __future__ import absolute_import, division, print_function
 
 from typing import TYPE_CHECKING, Any, Dict, Optional
 
-from six import StringIO
+from six import PY2, StringIO, string_types
 
 
 try:
@@ -222,7 +222,11 @@ class ConsoleRenderer(object):
                 + "] "
             )
 
+        # force event to str for compatibility with standard library
         event = event_dict.pop("event")
+        if not PY2 or not isinstance(event, string_types):
+            event = str(event)
+
         if event_dict:
             event = _pad(event, self._pad_event) + self._styles.reset + " "
         else: